% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADMM.q1.R
\name{ADMM.q1}
\alias{ADMM.q1}
\title{Subgroup detection in setting of regression with ADMM algorithm}
\usage{
ADMM.q1(Y, X, Z0, penalty, lam_vec, ga = 3.7, nu = 1, maxiter = 1000,
  ep = 5e-04)
}
\arguments{
\item{Y}{a n*1 vector, response.}

\item{X}{a n*p matrix, \code{p}-dimensional covariate.}

\item{Z0}{a n*1 matrix, treatment indicator.}

\item{lam_vec}{a series of candidate tuning parameter \code{lambda} for penalty function,
which is a vector.}

\item{ga}{a parameter for penalty function, default 3.7.}

\item{nu}{a tuning parameter for augmented Largranian objection function,
the default value is 1.}

\item{maxiter}{The maximum number of iterations. Defaults to 1000.}

\item{ep}{the absolute convergence tolerance.}

\item{peanlty}{\code{"SCAD"} or \code{"MCP"}.}
}
\value{
\itemize{
   \item \code{BIC} the values of BIC for each \code{lam_vec}.
   \item \code{lam_opt} the optimal \code{lambda} that have.
   the mimimal \code{BIC}.
   \item \code{beta_Est_lam} the regression coefficients matrix
   of covariates for different value of \code{lam_vec}, each column denotes a
   value of \code{lam_vec}.
   \item \code{theta_Est_lam}  the regression coefficients matrix of treatment indicator
   for each individuals, each column denotes a different value of \code{lam_vec}.
   \item \code{theta_est_opt} the regression coefficients of treatment indicator for
   optimal \code{lambda}.
   \item \code{beta_est_opt} the regression coefficients of covariates for optimal \code{lambda}.
}
}
\description{
The model is \deqn{  Y_i = X_i \beta + Z_i \theta_i.} This ADMM algorithm
can detect the subgroup effects of \eqn{\theta_i} and estimate the regression
coefficient \eqn{\beta} simultaneously.
}
\examples{
set.seed(1)
n = 200; p = 2
x = matrix(rnorm(n*p, mean = 2), nrow = n, ncol = p) # covariates
z = sample(c(0,1), n, T)    # treatment indicator
beta = c(-2, -3)            # regression coefficients of covariates
mu = c(-2, 2)               # regression coefficients of treatment indicator

y = x \%*\% beta + z * mu[1] + rnorm(n, sd = 0.5)
index = sample(c(0,1), n, T)
y[index == 1] = y[index == 1] + z[index == 1] * (mu[2] - mu[1])
lam_vec = seq(0,2,0.2)
result = ADMM.q1(y, x, z, penalty = 'SCAD', lam_vec = lam_vec)
result$beta_est_opt
unique( round(result$theta_est_opt, 1) )

# solution paths
py_mu = result$theta_Est_lam
plot(x=lam_vec, y=py_mu[1,], xlab=expression(lambda),
   ylab=expression(paste("Estimated value of  ", vartheta, sep=" ")),
   xlim=c(0,2), ylim=c(min(py_mu),max(py_mu)),
   main="Solution path for domain I", type='l')
for(j in 1:nrow(py_mu) ){
   lines(lam_vec, py_mu[j,], lty=j, col=j+1,lwd=2 )
}
abline(v=result$lam_opt, lty=3, col="lightblue",lwd =3)

}
\seealso{
\code{\link{ADMM}}
}
